requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.2
Pillow==8.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
